# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Delicious Dinner API",
        "description": "API realizada en Laravel",
        "contact": {
            "email": "ingisraeltrujillo@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/api/direccion_cliente": {
            "get": {
                "tags": [
                    "direccion_cliente"
                ],
                "summary": "Muestra todas las direcciones del cliente",
                "description": "Muestra todas las direcciones del cliente.",
                "operationId": "87b305ca02af9ac293686a74ed5138b9",
                "parameters": [
                    {
                        "name": "id_cliente",
                        "in": "query",
                        "description": "ID del cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Muestra todas las direcciones del cliente"
                    },
                    "500": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "tags": [
                    "direccion_cliente"
                ],
                "summary": "Crear una nueva dirección cliente",
                "operationId": "866721914211d1d3d9bd07ba53c4be21",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "id_cliente": {
                                                "type": "integer"
                                            },
                                            "pais": {
                                                "type": "string"
                                            },
                                            "ciudad": {
                                                "type": "string"
                                            },
                                            "barrio": {
                                                "type": "string"
                                            },
                                            "direccion_texto": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "id_cliente": 1,
                                    "pais": "Colombia",
                                    "ciudad": "Soledad",
                                    "barrio": "Centenario",
                                    "direccion_texto": "Calle 23B"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "CREATED"
                    }
                }
            }
        },
        "/api/direccion_cliente/{id}": {
            "get": {
                "tags": [
                    "direccion_cliente"
                ],
                "summary": "Mostrar la información de una dirección cliente",
                "operationId": "ebdeaf1a6b548de2f7c21274f80ca6dc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "put": {
                "tags": [
                    "direccion_cliente"
                ],
                "summary": "Actualizar una nueva dirección cliente",
                "operationId": "be16908e62700bb4899d430807d79d9c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "id_cliente": {
                                                "type": "integer"
                                            },
                                            "pais": {
                                                "type": "string"
                                            },
                                            "ciudad": {
                                                "type": "string"
                                            },
                                            "barrio": {
                                                "type": "string"
                                            },
                                            "direccion_texto": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "id_cliente": 1,
                                    "pais": "Colombia",
                                    "ciudad": "Soledad",
                                    "barrio": "Centenario",
                                    "direccion_texto": "Calle 23B"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "SUCCESS"
                    }
                }
            },
            "delete": {
                "tags": [
                    "direccion_cliente"
                ],
                "summary": "Eliminar una dirección cliente",
                "operationId": "568500bbdef8e624ff17829b37d4455b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESS",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "msg": "Se ha eliminado el cliente correctamente",
                                                "cliente": {
                                                    "id": 1,
                                                    "nombre": "CÉDULA DE CIUDADANÍA",
                                                    "codename": "CC",
                                                    "created_at": "2023-11-03T03:06:55.000000Z",
                                                    "updated_at": "2023-11-03T03:06:55.000000Z"
                                                }
                                            }
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}